Update BSPPolygon2.h
The BSPPolygon2::Finalize function passed *this as a non-const parameter and this->mEArray as a const parameter to the BSPTree2 constructor, which should have been a clue that the BSP implementation is badly done. A test data set was included in a bug report, and it led to an infinite loop in the BSPTree2 code because mEArray was modified via the *this object. A copy of mEArray is now passed as the second parameter of the BSPTree2 constructor.

The constructor also has a switch statement that classifies the intersection type of two edges. In the cases TRANSVERSE_POSITIVE and TRANSVERSE_NEGATIVE, it is possible that floating-point rounding error can cause the intersection point to be one of the segment endpoints. This must be trapped and handled separately. Before the change, the test data set triggered an assertion in SplitEdge for the user reporting the problem. I was unable to reproduces this, but compiled expressions using floating-point arithmetic can lead to different low-level code depending on the compiler, and the rounding behavior can vary. This is not just a matter of the low-level expression tree of operations, it can also be a matter of what precision registers are used and whether downcasts occur from a high-precision register (say, 80-bit) to a low-precision register