Replacing UniqueVerticesTriangles by UniqueVerticesSimplices.
Deprecated the class UniqueVerticesTriangles. Added a replacement, UniqueVerticesSimplices that works in any dimension 2 or larger. The class has been used for removing duplicated and unused vertices from triangle meshes and for generating meshes from a vertex-only based representation to one using an indexed representation. The generalized code applies to dimension 2 and will be used for computational geometry algorithms involving polygons (Boolean operations on polygons, for example). The deprecation is in the C++ 14 sense: the class is tagged with [[deprecated]]. It will be removed sometime in the near future.