Update IntrAlignedBox3Sphere3.h
The find-intersection query for moving spheres and oriented boxes calls the base-class protected query for moving spheres and aligned boxes. That call requires the center point to be in the first octant and makes adjustments accordingly. The find-intersection for the oriented boxes did not make this adjustment. I moved the transforming of the center point into the base-class query to fix the problem. The visual test harness for this code is the sample application MovingSpheresBoxes, where I apparently never moved the sphere into a position that exercised the bug.