Rewrites of ellipse/ellipsoid fitting code.
The old code ApprEllipse2 and ApprEllipsoid3 for fitting points with ellipses and ellipsoids were based on a nonlinear least-squares error function that was a sum of distances from points to the objects. The comments in those files described eigendecompositions of the M matrices but were in error, and the code did not correctly implement the mathematics. Moreover, ApprEllipsoid3 failed to initialize angle[2] before running the minimizer. The rewritten code uses a minimizer involving a 2-step gradient descent algorithm. Unit tests and end-to-end tests were added to the internal test suite to verify the algorithm and results.