Added comments about using Sylvester's criterion and removed a redundant test.
Added comments that Sylvester's criterion is used to test for positive definiteness. The loop on halving the root was limited to 8 iterations, but the number has been increased to allow rational arithmetic for which an unlimited loop could be infinite. The new limit is 2048, which will not be reached for floating-point types float and double. The ApprEllipse2 code had a redundant test for the determinant of the entire matrix, but it has now been removed.