Update DX11Engine.cpp
Swap chain creation in DX11 now has flip flags that are recommended. The DX11Engine still uses the old flags. Apparently, the internal changes to DX11 for the flip flags sometimes causes creation of a second ID3D11Context during a swap-chain Present call in order to create an ID3D11RenderTargetView and related objects. The DXGI debug layer indicates that the back buffer is unbound from the pipeline, which is not what I thought was happening with the old swap-chain creation flags. The second context appears to be created to patch the problem. However, the time of its destruction is nondeterministic. In some cases, the destruction occurs before the GTE application mDevice is destroyed, and the FinalRelease(mDevice) is successful. In other cases, the destruction occurs after the final release of mDevice, but an exception is thrown (by the GTE code) indicating the reference count for mDevice is positive. Finally, in many cases, the second context is never created. Until I understand why this behavior occurs, I have replaced the DX11::FinalRelease calls in DX11Engine::DestroyDevice with DX11::SafeRelease calls, the latter not tracking the reference counts and not throwing exceptions.