Modified TriangulateCDT to handle several classes of nonsimple polygons.
I redesigned and rewrote TriangulateCDT to support shared vertices, coincident vertex-edge configurations and coincident edge-edge configurations. These configurations are not supported by TriangulateEC. To avoid repackaging polygon trees in an application using both classes, I factored out the nested Tree classes into a new class PolygonTree. Objects of this class are inputs to the triangulators, and they are not modified internally. I also created a new class PolygonTreeEx that currently represents the output of TriangulateCDT. Eventually, this class will become an input-output parameter for both triangulators. Note that TriangulateCDT does not support polygon trees where two edges intersect at an edge-interior point of both edges and that intersection point is not in the input points to the triangulator. The problem is that the constrained Delaunay triangulation works correctly when the constrained edges do not intersect other than at shared-vertex endpoints. If one were to insert two edges into the Delaunay triangulation that have an edge-interior intersection, each edge interferes with the local re-triangulation of the other edge. I plan on adding support for such edges, but the task is of significant size and will take some time.