Update DX11Engine.cpp
The DX11Engine constructors that have a window handle input are designed for rendering to the window. These constructors create default render states that GTE wants, and these states are not all the default DX11 render states. The DX11Engine constructors that do not have a window handle input are designed for windowless applications using compute shaders. These constructors do not create default render states for GTE, so the default states are those of DX11. To support a windowless application that renders to an offscreen render target, and to allow that code to work whether windowed or windowless, I added creation of the default GTE render states to the compute-intended DX11Engine constructors. I also removed calls to DestroyDefaultObjects in the constructors because at the time of call, all the objects do not exist.