Create SWInterval.h
Added a new file for interval arithmetic. The implementation is software based and uses std::nextafter for determining the intervals. This avoids the floating-point hardware state changes when setting the rounding modes, and it avoids the problem with GCC that appears not to expose the ability to modify the floating-point environment.