Fixed the failure in MinimumVolumeBox3.h
I figured out why MinimumVolumeBox3D was broken. It turns out that the member function CreateCompactMesh has an implicit assumption that the vertices in the VETManifoldMesh::Vertex are ordered by the integer key. This assumption was satisifed when the vertex map was a std::map, but not when the vertex map was changed to a std::unordered_map. I restored the vertex map to be an unordered map and added code to the minimum-volume box to sort the vertices before creating the compact array of vertex adjacencies.