Fixed bug for line-sphere and line-ellipsoid find-intersection in tangent case.
The ray-sphere and segment-sphere queries were generating incorrect results when the linear component is tangent to the sphere. This occurred because result.parameter[1] is uninitialized in the line-sphere query for the tangent case, and the ray-sphere and segment-sphere code uses both result.parameter[] values for interval-interval intersection. The Result nested-class constructors now have all members initialized, and result.parameter[1] is set to result.parameter[0] in the tangent case. The same issues arise with ray-ellipsoid and segment-ellipsoid queries; they were fixed in the same manner.