Update BSNumber.h
The BSNumber constructors that take integer inputs need to compute the debug-supporting mValue first. The previous code modifies the inputs before computing mValue, which causes the sign to be positive when the inputs are negative. The
problem showed up only when GTE_BINARY_SCIENTIFIC_SHOW_DOUBLE was enabled for debugging and testing and does not affect the correctness of the actual BSNumber.