Update ParametricCurve.h
All but the first mAccumulatedLength elements are uninitialized, which causes GetTotalLength() to fail. The Release-build unit tests did not fail (via NaturalSplineCurve) using Microsoft Visual Studio because the compiler initializes the elements to zero. The compiler then optimizes away the first line of code in GetTotalLength() because it believes mAccumulatedLength.back() will return zero the first time GetTotalLength() is called. This hides the bug when using MSVS. Other compilers will most likely leave mAccumulatedLength uninitialized, in which case the bug shows up.