Update ConstrainedDelaunay2.h
The deprecated code in ConstrainedDelaunay2 (the first implementation in the header file) had a bug in Retriangulate that was introduced based on my blindly following a suggestion by the Microsoft Visual Studio code analysis tool. The description here is long but intended to let others know to be careful about the code analysis suggestions.

The Retriangulate function originally had a line of code to access the top of
the stack,
  auto i = stack[top--];
The Microsoft Visual Studio code analysis tool complained "'auto' doesn't deduce references. A possibly unintended copy is being made} and suggested modifying to a reference "auto&" (or "auto const&}. Without thinking about the message, I changed the code to
  auto const& i = stack[top--];
which is an error, because now I have a reference to the stack[top+1] element. Later in the function,
  stack[++top] = { i[0], isplit };
  stack[++top] = { isplit, i[1] };
In the first line of this code, the value "top" is incremented first and then the pair is copied to the stack. The problem is that "i" is the pair to which the left-hand side of the assignment refers. The write of "i[0]" to this location overwrites itself (no problem), but the write of "isplit" to this location overwrites "i[1]" (the error). The second line of code reads "i[1]", which was inadventently modified to "isplit". I have been burned several times by the faulty suggestions of the code analyzer and learned now to think about them, replacing the "auto" keyword by the actual type or ignoring them. The same problem had occurred in the new CDT code (the second implementation in the header file), but during unit testing I had found and fixed the problem. I had not thought to go back to the old code to see whether it had the same problem. My solution was to replace
  auto const& i = stack[top--];
by
  std::array<size_t, 2> i = stack[top--];