Ported the remaining Wild Magic 5 interface for IncrementalDelaunay2.
The Wild Magic 5 version of IncrementalDelaunay2 had support for computing the containing triangle for a query point. It also contained support for accessing the Delaunay triangles (not including triangles with at least one supervertex), for accessing the indices of triangles adjacent to a specified triangle, and for accessing the convex hull polygon of the Delaunay triangles. These have now been ported to the GTE version of IncrementalDelaunay2.