Update Hyperplane.h
The Hyperplane constructor whose input is a std::array of points uses singular value decomposition to build the plane through the points. The body has been replaced by a call to a function that is determined at compile time for dimension 3 and for other dimensions. This is a hack based on a bug report for the code executed on a QEMU/KVM virtual machine, which indicated the singular value decomposition was producing inaccurate results. I am unable to reproduce the problem on a non-virtual machine; the SVD works correctly for the dataset included in the bug report. I need to determine what the virtual machine is doing that causes such inaccurate results when using floating-point arithmetic. (The use of SVD in 3D is overkill anyway...)