Fixed GenerateMeshUV, removed the issue from the website Known Problems PDF.
The known-problems document mentioned that GenerateMeshUVs used to work on the GPU with HLSL but then stopped working, the returned texture coordinates all NaNs. It turns out that I had introduced the bug some time ago when factoring out the GPU code to GPUGenerateMeshUVs. The GenerateMeshUVs::Vertex data structure had 3 int members for use by the HLSL shader, where that shader declared a structured buffer array with int3 members. The data structure had 4 int members, the last one padding, for use by the GLSL shader. That shader declared a structured buffer array with ivec3 members, but GLSL pads each element to 16 bytes as if you had declared the members ivec4. The CPU padding was conditionally compiled. The factoring removed the conditional compilation, so now the GenerateMeshUVs::Vertex is 16 bytes regardless of using HLSL or GLSL. Unfortunately, I did not modify the HLSL code to use int4 instead of int3. This is now fixed. The document also mentions that when running the GLSL version, the PlanarMesh call to GetContainingTriangle was failing some of the time, unable to find a containing triangle. This occurs for GenerateMeshUVs<double> but not for GenerateMeshUVs<float>. The failure indicates fold-overs in the planar mesh. The double-precision arithmetic in the GLSL code appears not to be of the same precision as in that of the HLSL code.