Fixed various bugs in the intersection code.
The line-triangle find-intersection query needed to store the t-parameters as a pair of ordered numbers. This ensures derived classes can correctly compute interval-interval queries. The line-triangle test-intersection query determines on which side of the line the triangle vertices lie. The Boolean test for the related signed quantities was incorrectly implemented. The ray-triangle and segment-triangle queries call the line-triangle query. If that query reports an intersection, interval-interval queries are applied. It is possible that the interval queries report no-intersection, but the information was not propagated to the caller. While I was in there, I modified constructors to initialize \Code{Result} members (to avoid the code analysis complaints by Microsoft Visual Studio 2019).