Fixed a bug in IncrementalDelaunay2
The IncrementalDelaunay2 triangulation (with insertions and/or removals) uses a supertriangle approach. The construction of the insertion polygon can still include triangles sharing a supervertex, and in some cases the polygon is incorrect. Delaunay2 avoids the pitfalls of the supertriangle approach by not having one. To avoid a complete redesign, the IncrementalDelaunay2 code was modified to create the supertriangle with the input bounding rectangle included in the initial triangulation. The insertion polygon is now guaranteed to include only Delaunay triangles but not triangles having at least one supervertex. After all insertions and removals, if you do not want the input bounding rectangle as part of the triangulation, a new member function FinalizeTriangulation can be called to remove all triangles sharing an input rectangle vertex. After this call, you can no longer insert or remove points; that is, the triangulation really is final. I also fixed the version number for the header file and for the 3 files in the sample application. They all had 4.8.* when they should have had 5.8.*.