Relaxed restriction for line to have unit-length direction vector.
Modified the 2D line-triangle find-intersection query to allow for Line2 objects with direction vectors that are not unit length. This is necessary to support exact arithmetic in the forthcoming port of IntrTriangle3Triangle3 from Wild Magic to GTE. This query projects points from 3D to 2D and calls the IntrSegment2Triangle2 query.

NOTE: The modified IntrSegment2Triangle2 is a breaking change if you were accessing the Result.parameter[] values. In the previous code, the Result.parameter[] values were relative to the center-direction-extent form of the segment, where t in [-e,e] for extent e. The modified code does not convert the segment to center-direction-extent form in order to avoid floating-point rounding errors when computing the unit-length direction vector. Consequently, the modified code now has Result.parameter[] values relative to [0,1] where the segment is parameterized by p[0] + t (p[1] - p[0]) for t in [0,1].