Create IntrTriangle3Triangle3.h
Ported from Wild Magic to GTE the test-intersection and find-intersection queries for two stationary triangles in 3D. The template type can be floating-point, but rounding errors can cause misclassifications when two triangles are nearly touching. The template type can also be an arbitrary precision type, such as BSRational<UIntegerAP32>, and the intersection set (if any) is theoretically correct.