Reordered the coefficients of ApprQuadratic* to match that of QuadricSurface.
Reordered the coefficients of the fitted quadratic functions. Modified the coefficients argument to use std::array. This allows the array of output coefficients of ApprQuadratic3::operator() to be passed directly as an input to the QuadricSurface constructor. NOTE: The reordering will affect any of your applications that use the quadratic approximations.