Update SymmetricEigensolver3x3.h
The GetCosSin function was failing when the floating-point model was set to /fp:fast instead of the default model /fp:precise used by GTE. The function was simplified to avoid the problem. The PDF documentation "A Robust Eigensolver for 3x3 Symmetric Matrices" and the source code were revised to match each other. The PDF file is https://www.geometrictools.com/Documentation/RobustEigenSymmetric3x3.pdf