Added a test-intersection query for two finite cylinders.
Implemented the test-intersection query for two finite cylinders that is described in IntersectionOfCylinders.pdf. The original version of the PDF showed that generally computing a separating axis requires minimizing a convex function. No details were provided about implementing the minimizer. In an attempt to implement the query, it was clear that a deeper analysis was required. The latest version of the PDF has those details. Added a sample application to illustrate the ideas.