Update ApprCone3.h
Fixed a bug in the last blocks of code in function ApprCone3::ComputeInitialCone. The roots were supposed to be searched for the minimum, but the code for updating minItem did not update minError. The quartic roots are generally stored in increasing order except when a repeated root 0 is found (for the depressed quartic). The cone fitting should not lead to a root of 0, so the previous code worked. The least-squares fitting PDF at the Geometric Tools website was updated accordingly.