Update DistPoint3Circle3.h
The source code did not match the description for computing the distance robustly in the PDF entitled "Distance to Circles in 3D" (Section 2). It computed a closest point first, which was not robust, and then computed the distance using the input point and that closest point. The robustness problem occurs when the input point is nearly on the normal line C+tN, where C is the circle center and N is a unit-length normal to the plane of the circle. The distance computation now matches the PDF and is robust. The closest-point code was revised to be robust, and the PDF was expanded to describe the new code.