Modifications to some intersection queries.
Fixed a bug where the return value was not of type Result. The Result constructor had an input inEpsilon, but the caller of operator()(...) has no chance to specify this. The operator function now has a third parameter epsilon, defaulted to 0 to avoid breaking application code. The addition of 'explicit' to the Result constructor in IntrTriangle3Cylinder3 appears to be unnecessary because all branches the code can take have a return of a Result object. It is not possible to reach the end of the if-else statements without exiting, so there is no need for a return statement as the last line of the function. However, I modified the code to set result.intersect in each of the locks and added a single return statement as the last line in the function.  [Replacement of template Real by T is a flag to remind me I have ported and tested the code in the GTL development track. GTL also eliminates the annoyance of 'int' versus 'size_t' indexing. Inclusion of default-constructor braces is to avoid Visual Studio's code analyzer warning about potential uninitialized variables when in fact is is provable they are.]