Update TriangulateCDT.h
In FillRegion, accessing tkey from the front of regionQueue using a const reference is incorrect, because the next line of code is a pop of the queue which invalidates the memory, corrupting the value of tkey. A copy must be made of the front of the queue. Thanks to Aaron Barony for tracking down the problem.