Update SingularValueDecomposition.h
The bookkeeping for decomposing the bidiagonal matrix into a block matrix with bidiagonal blocks was not working correctly when a superdiagonal entry was zero. I decided to revisit every piece of the SVD code and updated the unit tests to ensure the code is working properly. The output of the SVD now produces only a singular-value matrix with diagonal elements occurring in descending order. The interfaces were modified slightly. (This is a significant revision of the code.)