Redesign of Logger system, initialization of class members
This is a major update. I removed the logging aspect of the "logger" system. Now all that Logger.h exposes are macros for assertions and errors that throw exceptions. This avoids duplicate symbol error messages from a linker when the Logger.h constructor is not inlined in all modules that are linked together. Also, I started the process of ensuring constructors initialize all members (using MSVS 2019 code analysis to find such issues). Many files have been modified as a result of the changes, although the LogAssert and LogError statements still exist (but only throw exceptions). To see a full list of changes, visit my Geometric Tools downloads page and click on the GteUpdateHistory.pdf file.