Added implementations for Sections 8.6 and 8.7 of the "Geometric Tools for Computer Graphics" book.
Added implementations of two algorithms in "Geometric Tools for Computer Graphics" (Sections 8.6 and 8.7) at user request. The implementation for the algorithm of Section 8.6 differs from that of the book--a simpler approach is used, described in the file comments. The book has pseudocode for the algorithm in Section 8.7, but there are numerous errors. I have added the corrections to the book correction page at the website. The approach I used is described in the file comments.