Fixed code for the intersection of a plane and a cylinder.
The source code failed when the plane normal and cylinder axis direction were parallel. The mathematical derivation in the document IntersectionPlaneCylinder.pdf was incorrect because it overlooked the singularity when computing the major axis of the ellipse of intersection. The PDF was rewritten and describes a different major axis construction. The new source code implements this. Moreover, it also now supports both finite and infinite cylinders (test-intersection and find-intersection queries).