Feature request: Fit cone to ellipse cross section and additional points.
Feature request to fit a cone to known elliptical cross sections and some
additional points (\Code{ApprCone3EllipseAndPoints}). If the cross sections
are provided as point samples approximately on the ellipses, the new code also
includes a class to extract points for each ellipse because the points might
be stored in a file in some unknown order (\Code{ApprCone3ExtractEllipses}).
3D ellipses are then fit to each subset of points in order to obtain the
ellipse input to \Code{ApprCone3EllipseAndPoints}. A PDF has been added to
the documentation to describe the fitting algorithm,
\Code{FitConeToEllipseAndPoints.pdf}.